+++ lib/src/controllers/types/mp_th2_edit_visualization_method.dart
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (C) 2023- Mapiah Ltda
+
+/// Rendering styles available for TH2 elements.
+enum MPTH2EditVisualizationMethod {
+  mapiahPlaceholder,
+  therionASF,
+  therionAUT,
+  therionBCRA,
+  therionNSS,
+  therionNZSS,
+  therionSBE,
+  therionSKBB,
+  therionSM,
+  therionUIS,
+}